Fixed null chapter

+++ app/src/main/java/com/inlurker/komiq/model/data/datamodel/Chapter.kt
@@ -6,7 +6,7 @@ import java.time.LocalDateTime
 
 data class Chapter(
     val id: String,
-    val volume: Int?,
+    val volume: Float?,
     val chapter: Float,
     val title: String,
     val publishAt: LocalDateTime,
